Blender addon for group node from script
Src layout modified to be used as folder for Blender script path to enable addon development from project. Copied antlr4 dependency to modules folder as a simple means of dependency management, as Blender does not provided a good way to manage pip-based dependencies.

Project is broken at the moment as a stable base to test parser development outside of Blender. Will address in a future commit.